SW-8378 Change Edit Org page to a modal
